# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * SFMA Community Platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface LoginInput {
  username: string;
  password: string;
}

export type MemberGrade = typeof MemberGrade[keyof typeof MemberGrade];


export const MemberGrade = {
  Fondateur_Suprême: 'Fondateur_Suprême',
  'Co-Fondateur': 'Co-Fondateur',
  Lord: 'Lord',
  Administrateur: 'Administrateur',
  Recruteur: 'Recruteur',
  Modérateur: 'Modérateur',
  Membre_Élite: 'Membre_Élite',
  Membre: 'Membre',
  Nouveau_Membre: 'Nouveau_Membre',
} as const;

export interface Member {
  id: string;
  username: string;
  grade: MemberGrade;
  gradeEmoji?: string;
  memberId: string;
  /** @nullable */
  country?: string | null;
  /** @nullable */
  whatsapp?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  /** @nullable */
  bio?: string | null;
  reputation: number;
  joinedAt: string;
  isBanned?: boolean;
  badges?: string[];
}

export interface AuthResult {
  member: Member;
  token: string;
}

export interface MemberUpdate {
  grade?: string;
  bio?: string;
  country?: string;
}

export interface AvatarInput {
  avatarUrl: string;
}

export type GroupWritePermission = typeof GroupWritePermission[keyof typeof GroupWritePermission];


export const GroupWritePermission = {
  all: 'all',
  lords_only: 'lords_only',
} as const;

export interface Group {
  id: string;
  name: string;
  emoji: string;
  description: string;
  writePermission: GroupWritePermission;
  order: number;
  memberCount?: number;
  /** @nullable */
  lastMessageAt?: string | null;
}

export type GroupInputWritePermission = typeof GroupInputWritePermission[keyof typeof GroupInputWritePermission];


export const GroupInputWritePermission = {
  all: 'all',
  lords_only: 'lords_only',
} as const;

export interface GroupInput {
  name: string;
  emoji: string;
  description: string;
  writePermission: GroupInputWritePermission;
  order?: number;
}

export type GroupUpdateWritePermission = typeof GroupUpdateWritePermission[keyof typeof GroupUpdateWritePermission];


export const GroupUpdateWritePermission = {
  all: 'all',
  lords_only: 'lords_only',
} as const;

export interface GroupUpdate {
  name?: string;
  emoji?: string;
  description?: string;
  writePermission?: GroupUpdateWritePermission;
  order?: number;
}

/**
 * @nullable
 */
export type MessageReplyTo = {
  id?: string;
  content?: string;
  authorUsername?: string;
} | null;

export interface Reaction {
  emoji: string;
  count: number;
  userIds: string[];
}

export interface Message {
  id: string;
  groupId: string;
  content: string;
  author: Member;
  /** @nullable */
  replyTo?: MessageReplyTo;
  reactions?: Reaction[];
  createdAt: string;
  mentions?: string[];
}

export interface MessageInput {
  content: string;
  /** @nullable */
  replyToId?: string | null;
  mentions?: string[];
}

export interface ReactionInput {
  emoji: string;
}

export type ApplicationStatus = typeof ApplicationStatus[keyof typeof ApplicationStatus];


export const ApplicationStatus = {
  pending: 'pending',
  accepted: 'accepted',
  rejected: 'rejected',
} as const;

export interface Application {
  id: string;
  pseudo: string;
  age: number;
  country: string;
  whatsapp: string;
  /** @nullable */
  previousClan?: string | null;
  /** @nullable */
  previousClanLeaveReason?: string | null;
  availability?: string;
  reason: string;
  valueContribution?: string;
  respectAnswer?: string;
  conflictAnswer?: string;
  status: ApplicationStatus;
  submittedAt: string;
  /** @nullable */
  reviewedAt?: string | null;
  /** @nullable */
  reviewedBy?: string | null;
}

export interface ApplicationInput {
  pseudo: string;
  age: number;
  country: string;
  whatsapp: string;
  previousClan?: string;
  previousClanLeaveReason?: string;
  availability: string;
  reason: string;
  valueContribution: string;
  respectAnswer: string;
  conflictAnswer: string;
  desiredPassword: string;
}

export interface PollOption {
  id: string;
  text: string;
  votes: number;
  percentage?: number;
}

export interface Poll {
  id: string;
  question: string;
  options: PollOption[];
  createdAt: string;
  createdBy: string;
  /** @nullable */
  closedAt?: string | null;
  totalVotes?: number;
  /** @nullable */
  userVotedOption?: string | null;
}

export interface PollInput {
  question: string;
  options: string[];
}

export interface VoteInput {
  optionId: string;
}

export interface Event {
  id: string;
  title: string;
  description: string;
  /** @nullable */
  location?: string | null;
  eventDate: string;
  createdBy: string;
  participantCount: number;
  isParticipating?: boolean;
  participants?: string[];
}

export interface EventInput {
  title: string;
  description: string;
  location?: string;
  eventDate: string;
}

export type GalleryPhotoStatus = typeof GalleryPhotoStatus[keyof typeof GalleryPhotoStatus];


export const GalleryPhotoStatus = {
  pending: 'pending',
  approved: 'approved',
  rejected: 'rejected',
} as const;

export interface GalleryPhoto {
  id: string;
  url: string;
  /** @nullable */
  caption?: string | null;
  submittedBy: string;
  status: GalleryPhotoStatus;
  submittedAt: string;
  /** @nullable */
  approvedAt?: string | null;
}

export interface PhotoInput {
  url: string;
  caption?: string;
}

export interface SiteSettings {
  siteName: string;
  slogan: string;
  description: string;
  /** @nullable */
  logoUrl?: string | null;
  /** @nullable */
  backgroundImageUrl?: string | null;
  primaryColor: string;
  accentColor: string;
  backgroundColor: string;
  fontFamily?: string;
  /** @nullable */
  whatsappGroupLink?: string | null;
  /** @nullable */
  whatsappGroupPhoto?: string | null;
}

export interface SiteSettingsUpdate {
  siteName?: string;
  slogan?: string;
  description?: string;
  logoUrl?: string;
  backgroundImageUrl?: string;
  primaryColor?: string;
  accentColor?: string;
  backgroundColor?: string;
  fontFamily?: string;
  whatsappGroupLink?: string;
  whatsappGroupPhoto?: string;
}

export type CommunityStatsGradeDistributionItem = {
  grade: string;
  count: number;
};

export interface CommunityStats {
  totalMembers: number;
  activeMembers: number;
  totalMessages: number;
  totalEvents: number;
  pendingApplications?: number;
  gradeDistribution: CommunityStatsGradeDistributionItem[];
  recentJoins: Member[];
}

export interface AuditEntry {
  id: string;
  actorUsername: string;
  action: string;
  targetType: string;
  /** @nullable */
  targetId?: string | null;
  /** @nullable */
  targetName?: string | null;
  /** @nullable */
  details?: string | null;
  timestamp: string;
}