# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: SFMA Community Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Login with username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResult"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: Current user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Member"
        "401":
          description: Not authenticated
...
  /groups:
    get:
      operationId: listGroups
      tags: [groups]
      summary: List all groups
      responses:
        "200":
          description: Groups list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Group"

    post:
      operationId: createGroup
      tags: [groups]
      summary: Create a group (Lords only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupInput"
      responses:
        "201":
          description: Created group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"
  /groups/{id}:
    get:
      operationId: getGroup
      tags: [groups]
      summary: Get group details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"

    patch:
      operationId: updateGroup
      tags: [groups]
      summary: Update group (Lords only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/GroupUpdate"
      responses:
        "200":
          description: Updated group
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Group"

    delete:
      operationId: deleteGroup
      tags: [groups]
      summary: Delete group (Lords only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Deleted
...
  /applications:
    get:
      operationId: listApplications
      tags: [applications]
      summary: List applications (Lords only)
      responses:
        "200":
          description: Applications
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Application"

    post:
      operationId: submitApplication
      tags: [applications]
      summary: Submit a recruitment application
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "201":
          description: Application submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/accept:
    post:
      operationId: acceptApplication
      tags: [applications]
      summary: Accept an application (Lords only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Accepted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
  /applications/{id}/reject:
    post:
      operationId: rejectApplication
      tags: [applications]
      summary: Reject an application (Lords only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Rejected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Application"
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List community events
      responses:
        "200":
          description: Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Event"

    post:
      operationId: createEvent
      tags: [events]
      summary: Create event (Lords only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}/rsvp:
    post:
      operationId: rsvpEvent
      tags: [events]
      summary: RSVP to an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: RSVP toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /gallery:
    get:
      operationId: listGallery
      tags: [gallery]
      summary: List approved gallery photos
      responses:
        "200":
          description: Gallery photos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GalleryPhoto"

    post:
      operationId: submitPhoto
      tags: [gallery]
      summary: Submit a photo for approval
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PhotoInput"
      responses:
        "201":
          description: Photo submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
  /gallery/{id}/approve:
    post:
      operationId: approvePhoto
      tags: [gallery]
      summary: Approve a gallery photo (Lords only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Approved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GalleryPhoto"
...
components:
  schemas:
...
    LoginInput:
      type: object
      required: [username, password]
      properties:
        username:
          type: string
        password:
          type: string
    AuthResult:
      type: object
      required: [member, token]
      properties:
        member:
          $ref: "#/components/schemas/Member"
        token:
          type: string
    Member:
      type: object
      required: [id, username, grade, memberId, joinedAt, reputation]
      properties:
        id:
          type: string
        username:
          type: string
        grade:
          type: string
          enum: [Fondateur_Suprême, Co-Fondateur, Lord, Administrateur, Recruteur, Modérateur, Membre_Élite, Membre, Nouveau_Membre]
        gradeEmoji:
          type: string
        memberId:
          type: string
        country:
          type: ["string", "null"]
        whatsapp:
          type: ["string", "null"]
        avatarUrl:
          type: ["string", "null"]
        bio:
          type: ["string", "null"]
        reputation:
          type: integer
        joinedAt:
          type: string
        isBanned:
          type: boolean
        badges:
          type: array
          items:
            type: string
...
    Group:
      type: object
      required: [id, name, emoji, description, writePermission, order]
      properties:
        id:
          type: string
        name:
          type: string
        emoji:
          type: string
        description:
          type: string
        writePermission:
          type: string
          enum: [all, lords_only]
        order:
          type: integer
        memberCount:
          type: integer
        lastMessageAt:
          type: ["string", "null"]
    GroupInput:
      type: object
      required: [name, emoji, description, writePermission]
      properties:
        name:
          type: string
        emoji:
          type: string
        description:
          type: string
        writePermission:
          type: string
          enum: [all, lords_only]
        order:
          type: integer
    GroupUpdate:
      type: object
      properties:
        name:
          type: string
        emoji:
          type: string
        description:
          type: string
        writePermission:
          type: string
          enum: [all, lords_only]
        order:
          type: integer
...
    Application:
      type: object
      required: [id, pseudo, age, country, whatsapp, reason, status, submittedAt]
      properties:
        id:
          type: string
        pseudo:
          type: string
        age:
          type: integer
        country:
          type: string
        whatsapp:
          type: string
        previousClan:
          type: ["string", "null"]
        previousClanLeaveReason:
          type: ["string", "null"]
        availability:
          type: string
        reason:
          type: string
        valueContribution:
          type: string
        respectAnswer:
          type: string
        conflictAnswer:
          type: string
        status:
          type: string
          enum: [pending, accepted, rejected]
        submittedAt:
          type: string
        reviewedAt:
          type: ["string", "null"]
        reviewedBy:
          type: ["string", "null"]
    ApplicationInput:
      type: object
      required: [pseudo, age, country, whatsapp, availability, reason, valueContribution, respectAnswer, conflictAnswer, desiredPassword]
      properties:
        pseudo:
          type: string
        age:
          type: integer
        country:
          type: string
        whatsapp:
          type: string
        previousClan:
          type: string
        previousClanLeaveReason:
          type: string
        availability:
          type: string
        reason:
          type: string
        valueContribution:
          type: string
        respectAnswer:
          type: string
        conflictAnswer:
          type: string
        desiredPassword:
          type: string
...
    Event:
      type: object
      required: [id, title, description, eventDate, createdBy, participantCount]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        location:
          type: ["string", "null"]
        eventDate:
          type: string
        createdBy:
          type: string
        participantCount:
          type: integer
        isParticipating:
          type: boolean
        participants:
          type: array
          items:
            type: string
    EventInput:
      type: object
      required: [title, description, eventDate]
      properties:
        title:
          type: string
        description:
          type: string
        location:
          type: string
        eventDate:
          type: string
    GalleryPhoto:
      type: object
      required: [id, url, submittedBy, status, submittedAt]
      properties:
        id:
          type: string
        url:
          type: string
        caption:
          type: ["string", "null"]
        submittedBy:
          type: string
        status:
          type: string
          enum: [pending, approved, rejected]
        submittedAt:
          type: string
        approvedAt:
          type: ["string", "null"]
    PhotoInput:
      type: object
      required: [url]
      properties:
        url:
          type: string
        caption:
          type: string
...